requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==2024.2
cryptography==43.0.0
cssselect2==0.7.0
Django==4.2.14
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-extensions==3.2.3
django-feather==0.3.0
django-formtools==2.5.1
django-libsass==0.9
django-livereload-server==0.5.1
django_compressor==4.5
djangorestframework==3.12.4
flake8==3.9.2
fonttools==4.53.1
gunicorn==22.0.0
html5lib==1.1
idna==3.7
libsass==0.23.0
lxml==5.3.0
mccabe==0.6.1
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pycparser==2.22
pydyf==0.11.0
pyflakes==2.3.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
rcssmin==1.1.2
reportlab==4.2.2
requests==2.32.3
rjsmin==1.2.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
django-formtools
flake8
channels
